base: Fix compilation errors
Mostly caused by inconsistencies between declaration/definition.

+++ src/base/p_scatter.c
@@ -2,7 +2,7 @@
 #include "pal_base.h"
 #include "pal_base_private.h"
 
-int p_scatter(void *src, size_t nsrc, size_t ndst, int flags, void **dstlist)
+ssize_t p_scatter(void *src, size_t nsrc, size_t ndst, int flags, void **dstlist)
 {
 
     /*PLACE CODE HERE*/
